Update after code review

+++ core/src/wallet/bitcoin/api_impl/BitcoinLikeTransactionApi.cpp
@@ -321,15 +321,15 @@ namespace ledger {
             return estimateSize(fixedSize, inputCount, outputsSize, outputsSize, keychainEngine);
         }
 
-        int64_t BitcoinLikeTransactionApi::computeWorthlessUtxoValue(const api::Currency &currency, const std::string &keychainEngine, const std::vector<std::shared_ptr<api::BigInt>> &fees) {
-            // Compute cost of change
+        int32_t BitcoinLikeTransactionApi::computeWorthlessUtxoValue(const api::Currency &currency, const std::string &keychainEngine, const std::vector<std::shared_ptr<api::BigInt>> &fees) {
+            // Size of empty transaction (0 input 0 output)
             auto const fixedSize         = BitcoinLikeTransactionApi::estimateSize(0,
                                                                                    0,
                                                                                    currency,
                                                                                    keychainEngine);
 
             // Size 1 signed UTXO (signed input)
-            const int64_t signedUTXOSize = BitcoinLikeTransactionApi::estimateSize(1,
+            const int32_t signedUTXOSize = BitcoinLikeTransactionApi::estimateSize(1,
                                                                                    0,
                                                                                    currency,
                                                                                    keychainEngine)
